(PC-13382)[PRO] refactor(offer): extract view-in-app link in a generic component
