+++ pro/src/components/pages/Offers/Offer/DisplayOfferInAppLink/index.ts
@@ -1,0 +1,1 @@
+export * from './DisplayOfferInAppLink'
