+++ pro/src/new_components/DisplayInAppLink/index.ts
@@ -1,0 +1,1 @@
+export * from './DisplayInAppLink'
